api/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
SQLAlchemy-serializer
beautifulsoup4
pystan
statsmodels
yahoo-finance
yfinance
xgboost
pmdarima
